+++ features/admin.applications.v1/models/application.ts
@@ -847,7 +847,9 @@ export enum ApplicationTemplateIdTypes {
     MOBILE_APPLICATION = "mobile-application",
     M2M_APPLICATION = "m2m-application",
     CUSTOM_APPLICATION = "custom-application",
-    MCP_CLIENT_APPLICATION = "mcp-client-application"
+    MCP_CLIENT_APPLICATION = "mcp-client-application",
+    REACT_APPLICATION = "react-application",
+    NEXT_JS_APPLICATION = "nextjs-application"
 }
 
 /**