+++ features/admin.server-configurations.v1/pages/connector-listing-page.tsx
@@ -36,7 +36,6 @@ import { Placeholder, Ref } from "semantic-ui-react";
 import { getConnectorCategories, getConnectorCategory } from "../api";
 import GovernanceConnectorCategoriesGrid from "../components/governance-connector-grid";
 import { ServerConfigurationsConstants } from "../constants";
-import useRegistrationFlowBuilderConnector from "../hooks/use-registration-flow-builder-connector";
 import {
     ConnectorOverrideConfig,
     GovernanceConnectorCategoryInterface,
@@ -71,8 +70,6 @@ export const ConnectorListingPage: FunctionComponent<ConnectorListingPageInterfa
     const { t } = useTranslation();
     const { UIConfig } = useUIConfig();
 
-    const registrationFlowBuilderConnector: unknown = useRegistrationFlowBuilderConnector();
-
     const featureConfig: FeatureConfigInterface = useSelector((state: AppState) => state.config.ui.features);
     const allowedScopes: string = useSelector((state: AppState) => state?.auth?.allowedScopes);
     const isPasswordInputValidationEnabled: boolean = useSelector((state: AppState) =>
@@ -95,11 +92,7 @@ export const ConnectorListingPage: FunctionComponent<ConnectorListingPageInterfa
     );
 
     const predefinedCategories: any = useMemo(() => {
-        let originalConnectors: Array<any> = GovernanceConnectorUtils.getCombinedPredefinedConnectorCategories();
-
-        originalConnectors = GovernanceConnectorUtils.addAdditionalConnectors(originalConnectors, [
-            registrationFlowBuilderConnector
-        ]);
+        const originalConnectors: Array<any> = GovernanceConnectorUtils.getCombinedPredefinedConnectorCategories();
 
         const refinedConnectorCategories: Array<any> = [];
 
@@ -136,7 +129,7 @@ export const ConnectorListingPage: FunctionComponent<ConnectorListingPageInterfa
         }
 
         return refinedConnectorCategories;
-    }, [ featureConfig, UIConfig, allowedScopes, registrationFlowBuilderConnector ]);
+    }, [ featureConfig, UIConfig, allowedScopes ]);
 
     const [
         dynamicConnectorCategories,