+++ features/admin.template-core.v1/models/templates.ts
@@ -97,7 +97,7 @@ export interface ExtensionTemplateCategoryInterface {
     /**
      * Display name of the template category.
      */
-    displayName: string;
+    displayName?: string;
     /**
      * Description of the template category.
      */